Extract comment parser
